refactor: remove cycle from PaymentMethod
