+++ src/components/Employee/PaymentMethod/BankAccountEdit.tsx
@@ -1,29 +1,9 @@
 import { useFormContext } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
-import {
-  object,
-  pipe,
-  string,
-  nonEmpty,
-  regex,
-  picklist,
-  type InferNonNullableInput,
-  literal,
-} from 'valibot'
-import { usePaymentMethod } from '@/components/Employee/PaymentMethod/PaymentMethod'
+import type { BankAccountInputs } from './BankAccount'
+import { usePaymentMethod } from './usePaymentMethod'
 import { PAYMENT_METHODS } from '@/components/Employee/PaymentMethod/PaymentTypeForm'
 import { RadioGroup, TextField } from '@/components/Common'
-
-export const BankAccountSchema = object({
-  name: pipe(string(), nonEmpty('f')),
-  routingNumber: pipe(string(), regex(/^[0-9]{9}$/)),
-  accountNumber: pipe(string(), regex(/^[0-9]{9,18}$/)),
-  accountType: picklist(['Checking', 'Savings']),
-  hasBankPayload: literal(true),
-})
-
-export type BankAccountInputs = InferNonNullableInput<typeof BankAccountSchema>
-export type BankAccountDefaults = InferNonNullableInput<typeof BankAccountSchema>
 
 export const BankAccountForm = () => {
   const { mode, watchedType } = usePaymentMethod()
