+++ src/components/Employee/PaymentMethod/PaymentTypeForm.tsx
@@ -2,7 +2,7 @@ import { Text } from 'react-aria-components'
 import { useFormContext } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
 import * as v from 'valibot'
-import { usePaymentMethod } from '@/components/Employee/PaymentMethod/PaymentMethod'
+import { usePaymentMethod } from './usePaymentMethod'
 import { RadioGroup } from '@/components/Common'
 
 export const PaymentTypeSchema = v.object({
